+++ packages/atomic/src/components/search-box-suggestions/atomic-search-box-instant-results/atomic-search-box-instant-results.tsx
@@ -1,4 +1,4 @@
-import {Component, Element, State, h} from '@stencil/core';
+import {Component, Element, State, h, Prop} from '@stencil/core';
 import {buildInstantResults, InstantResults, Result} from '@coveo/headless';
 
 import {
@@ -7,6 +7,16 @@ import {
   SearchBoxSuggestions,
   SearchBoxSuggestionsBindings,
 } from '../suggestions-common';
+import {cleanUpString} from '../../../utils/string-utils';
+import {
+  BaseResultList,
+  ResultListCommon,
+} from '../../result-lists/result-list-common';
+import {
+  ResultDisplayDensity,
+  ResultDisplayImageSize,
+  ResultDisplayLayout,
+} from '../../atomic-result/atomic-result-display-options';
 
 /**
  * The `atomic-search-box-instant-results` component can be added as a child of an `atomic-search-box` component, allowing for the configuration of instant results behavior.
@@ -16,15 +26,30 @@ import {
   tag: 'atomic-search-box-instant-results',
   shadow: true,
 })
-export class AtomicSearchBoxRecentQueries {
-  private bindings!: SearchBoxSuggestionsBindings;
+export class AtomicSearchBoxInstantResults implements BaseResultList {
+  public bindings!: SearchBoxSuggestionsBindings;
 
-  @Element() private host!: HTMLElement;
+  @Element() public host!: HTMLElement;
 
   @State() public error!: Error;
+  @State() public templateHasError = false;
   private instantResults!: InstantResults;
 
   private results: Result[] = [];
+  public resultListCommon!: ResultListCommon;
+  /**
+   * The desired layout to use when displaying results. Layouts affect how many results to display per row and how visually distinct they are from each other.
+   */
+  @Prop({reflect: true}) display: ResultDisplayLayout = 'list';
+
+  /**
+   * The spacing of various elements in the result list, including the gap between results, the gap between parts of a result, and the font sizes of different parts in a result.
+   */
+  @Prop({reflect: true}) density: ResultDisplayDensity = 'normal';
+  /**
+   * The expected size of the image displayed in the results.
+   */
+  @Prop({reflect: true}) imageSize: ResultDisplayImageSize = 'icon';
 
   componentWillLoad() {
     try {
@@ -38,13 +63,27 @@ export class AtomicSearchBoxRecentQueries {
   }
 
   private renderItems(): SearchBoxSuggestionItem[] {
+    if (!this.bindings.suggestedQuery()) {
+      return [];
+    }
     const results = this.instantResults.state.results.length
       ? this.instantResults.state.results
       : this.results;
-    return results.map((result, i) => ({
-      key: `instant-result-${i}`,
+    return results.map((result: Result) => ({
+      key: `instant-result-${cleanUpString(result.uniqueId)}`,
       query: '',
-      content: <div class="flex items-center break-all">{result.title}</div>,
+      content: (
+        <atomic-result
+          key={`instant-results-${result.uniqueId}`}
+          part="outline"
+          result={result}
+          engine={this.bindings.engine}
+          display={this.display}
+          density={this.density}
+          imageSize={this.imageSize}
+          content={this.resultListCommon.getContentOfResultTemplate(result)}
+        ></atomic-result>
+      ),
       onSelect: () => {
         // TODO: ADD LOGS?
       },
@@ -58,6 +97,15 @@ export class AtomicSearchBoxRecentQueries {
       },
     });
 
+    this.resultListCommon = new ResultListCommon({
+      host: this.host,
+      bindings: this.bindings,
+      templateElements: this.host.querySelectorAll('atomic-result-template'),
+      onReady: () => {},
+      onError: () => {
+        this.templateHasError = true;
+      },
+    });
     return {
       position: Array.from(this.host.parentNode!.children).indexOf(this.host),
       panel: 'right',