+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.57.1](https://github.com/coveo/ui-kit/compare/headless-react@0.57.0...headless-react@0.57.1) (2022-06-09)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.57.0](https://github.com/coveo/ui-kit/compare/headless-react@0.56.2...headless-react@0.57.0) (2022-06-09)
+
+
+### Features
+
+* **atomic:** extract result list interface ([#2064](https://github.com/coveo/ui-kit/issues/2064)) ([4e0054d](https://github.com/coveo/ui-kit/commit/4e0054d4229fe594b8ab44954cf6941cf3745336))
+
+
+
+
+
+## [0.56.2](https://github.com/coveo/ui-kit/compare/headless-react@0.56.1...headless-react@0.56.2) (2022-06-08)
+
+
+### Bug Fixes
+
+* **atomic:** fix issue with changing panels while results load ([#2060](https://github.com/coveo/ui-kit/issues/2060)) ([d10a8e9](https://github.com/coveo/ui-kit/commit/d10a8e90d170bc23dda19b3dd59aa9bf70f61a70))
+
+
+
+
+
+## [0.56.1](https://github.com/coveo/ui-kit/compare/headless-react@0.56.0...headless-react@0.56.1) (2022-06-08)
+
+
+### Bug Fixes
+
+* **headless:** triggered notify-trigger subscriber when notifications are removed ([#2055](https://github.com/coveo/ui-kit/issues/2055)) ([8061479](https://github.com/coveo/ui-kit/commit/80614791c447451707bef8f52d78e9fb5de7c37e))
+
+
+
+
+
+# [0.56.0](https://github.com/coveo/ui-kit/compare/headless-react@0.55.1...headless-react@0.56.0) (2022-06-07)
+
+
+### Features
+
+* **atomic:** hide facets on "no results" & "error" status (search layout) ([#2053](https://github.com/coveo/ui-kit/issues/2053)) ([5d75cd2](https://github.com/coveo/ui-kit/commit/5d75cd2e6f73777b44dda0e20a5c01b001315e8b))
+
+
+
+
+
+## [0.55.1](https://github.com/coveo/ui-kit/compare/headless-react@0.55.0...headless-react@0.55.1) (2022-06-06)
+
+
+### Bug Fixes
+
+* **atomic:** made facet values focus when pressing show more/show less ([#2043](https://github.com/coveo/ui-kit/issues/2043)) ([1e77f3e](https://github.com/coveo/ui-kit/commit/1e77f3e8bbd23c97f240e11685fcabb5029bb50d))
+
+
+
+
+
+# [0.55.0](https://github.com/coveo/ui-kit/compare/headless-react@0.54.1...headless-react@0.55.0) (2022-06-06)
+
+
+### Features
+
+* **headless:** add insight "get interface" action ([#2036](https://github.com/coveo/ui-kit/issues/2036)) ([942dfe8](https://github.com/coveo/ui-kit/commit/942dfe8d72de29ad4b2ac619dc2f9704148325d3))
+
+
+
+
+
+## [0.54.1](https://github.com/coveo/ui-kit/compare/headless-react@0.54.0...headless-react@0.54.1) (2022-06-03)
+
+
+### Bug Fixes
+
+* **atomic:** add getAnimations() check ([#2040](https://github.com/coveo/ui-kit/issues/2040)) ([47161a3](https://github.com/coveo/ui-kit/commit/47161a3ec2ef14b9c454a04bad6abbdd21f0ef0d))
+
+
+
+
+
+# [0.54.0](https://github.com/coveo/ui-kit/compare/headless-react@0.53.1...headless-react@0.54.0) (2022-06-02)
+
+
+### Features
+
+* **atomic:** change dayjs language when i18n language changes ([#2033](https://github.com/coveo/ui-kit/issues/2033)) ([f6093cd](https://github.com/coveo/ui-kit/commit/f6093cd9f575a26c821b7bf2698bdcf429871c12))
+
+
+
+
+
 ## [0.53.1](https://github.com/coveo/ui-kit/compare/headless-react@0.53.0...headless-react@0.53.1) (2022-06-01)
 
 **Note:** Version bump only for package headless-react